refactro: useFilter
