+++ apps/web/src/views/GaugesVoting/components/GaugesFilter/FilterFieldByType.tsx
@@ -1,8 +1,9 @@
 import { useTranslation } from '@pancakeswap/localization'
 import { AutoColumn, Button, FlexGap, Text, useMatchBreakpoints } from '@pancakeswap/uikit'
-import React from 'react'
+import React, { useState } from 'react'
 import styled from 'styled-components'
-import { OptionsType } from './type'
+import { FilterModal } from './FilterModal'
+import { Filter, FilterValue, OptionsType } from './type'
 
 const FilterButton = styled(Button)`
   height: 35px;
@@ -12,29 +13,40 @@ const FilterButton = styled(Button)`
 `
 
 type FilterButtonGroupProps = {
-  onFilterChange: (filter: OptionsType) => void
+  onFilterChange: (type: OptionsType, value: FilterValue) => void
+  value: Filter
 }
 
-export const FilterFieldByType: React.FC<FilterButtonGroupProps> = ({ onFilterChange }) => {
+export const FilterFieldByType: React.FC<FilterButtonGroupProps> = ({ onFilterChange, value }) => {
   const { t } = useTranslation()
+  const [option, setOption] = useState<OptionsType | null>(null)
   const { isMobile } = useMatchBreakpoints()
 
   return (
-    <AutoColumn gap="4px">
-      <Text fontSize={12} fontWeight={600} color="textSubtle" textTransform="uppercase">
-        {t('filter')}
-      </Text>
-      <FlexGap gap={isMobile ? '0' : '10px'} justifyContent={isMobile ? 'space-between' : 'flex-start'}>
-        <FilterButton variant="light" onClick={() => onFilterChange(OptionsType.ByChain)}>
-          {t('Chain')}
-        </FilterButton>
-        <FilterButton variant="light" onClick={() => onFilterChange(OptionsType.ByFeeTier)}>
-          {t('Fee Tier')}
-        </FilterButton>
-        <FilterButton variant="light" onClick={() => onFilterChange(OptionsType.ByType)}>
-          {t('Type')}
-        </FilterButton>
-      </FlexGap>
-    </AutoColumn>
+    <>
+      <AutoColumn gap="4px">
+        <Text fontSize={12} fontWeight={600} color="textSubtle" textTransform="uppercase">
+          {t('filter')}
+        </Text>
+        <FlexGap gap={isMobile ? '0' : '10px'} justifyContent={isMobile ? 'space-between' : 'flex-start'}>
+          <FilterButton variant="light" onClick={() => setOption(OptionsType.ByChain)}>
+            {t('Chain')}
+          </FilterButton>
+          <FilterButton variant="light" onClick={() => setOption(OptionsType.ByFeeTier)}>
+            {t('Fee Tier')}
+          </FilterButton>
+          <FilterButton variant="light" onClick={() => setOption(OptionsType.ByType)}>
+            {t('Type')}
+          </FilterButton>
+        </FlexGap>
+      </AutoColumn>
+      <FilterModal
+        isOpen={Boolean(option)}
+        onDismiss={() => setOption(null)}
+        type={option}
+        options={value}
+        onChange={onFilterChange}
+      />
+    </>
   )
 }
