+++ apps/web/src/views/GaugesVoting/components/Table/AddGauge/AddGaugeModal.tsx
@@ -1,5 +1,3 @@
-import { Gauge, GaugeType } from '@pancakeswap/gauges'
-import { useDebounce } from '@pancakeswap/hooks'
 import { useTranslation } from '@pancakeswap/localization'
 import {
   AutoColumn,
@@ -16,19 +14,11 @@ import {
   Text,
   useMatchBreakpoints,
 } from '@pancakeswap/uikit'
-import { useEffect, useMemo, useState } from 'react'
 import styled from 'styled-components'
+import { useFilter } from 'views/GaugesVoting/hooks/useFilter'
 import { useGauges } from 'views/GaugesVoting/hooks/useGauges'
 import { useGaugesTotalWeight } from 'views/GaugesVoting/hooks/useGaugesTotalWeight'
-import {
-  Filter,
-  FilterFieldByType,
-  FilterFieldInput,
-  FilterModal,
-  FilterValue,
-  Gauges,
-  OptionsType,
-} from '../../GaugesFilter'
+import { FilterFieldByType, FilterFieldInput } from '../../GaugesFilter'
 import { GaugesList, GaugesTable } from '../GaugesTable'
 import { THeader, TRow } from '../styled'
 
@@ -51,71 +41,7 @@ export const AddGaugeModal = ({ isOpen, onDismiss, selectRows, onGaugeAdd }) => 
   const { isDesktop, isMobile } = useMatchBreakpoints()
   const totalGaugesWeight = useGaugesTotalWeight()
   const { data: gauges } = useGauges()
-  const [searchText, setSearchText] = useState<string>('')
-  const debouncedSearchText = useDebounce(searchText, 800)
-  const [option, setOption] = useState<OptionsType | null>(null)
-  const [filter, setFilter] = useState<Filter>({
-    byChain: [],
-    byFeeTier: [],
-    byType: [],
-  })
-
-  const filterRows = useMemo(() => {
-    if (!gauges || !gauges.length) return []
-    const { byChain, byFeeTier, byType } = filter
-    let rows: Gauge[] = gauges
-
-    if (debouncedSearchText?.length > 0) {
-      rows = gauges.filter((gauge) => {
-        return gauge.pairName.toLowerCase().includes(debouncedSearchText.toLowerCase())
-      })
-    }
-
-    if (byChain.length || byFeeTier.length || byType.length) {
-      rows = rows?.filter((gauge: Gauge) => {
-        const feeTier = gauge.type === GaugeType.V3 ? gauge?.feeTier : undefined
-        const chain = gauge.chainId
-        const boosted = gauge.boostMultiplier > 100n
-        const capped = gauge.maxVoteCap > 0n
-        const types = [boosted ? Gauges.Boosted : Gauges.Regular]
-        if (capped) {
-          types.push(Gauges.Capped)
-        }
-        return (
-          (byChain.length === 0 || (chain && byChain.includes(chain))) &&
-          (byFeeTier.length === 0 || (feeTier && byFeeTier.includes(feeTier))) &&
-          (byType.length === 0 || byType.some((bt) => types.includes(bt)))
-        )
-      })
-    }
-
-    return rows
-  }, [filter, gauges, debouncedSearchText])
-
-  const onFilterChange = (type: OptionsType, value: FilterValue) => {
-    const opts = filter[type] as Array<unknown>
-
-    // select all
-    if (Array.isArray(value)) {
-      setFilter((prev) => ({
-        ...prev,
-        [type]: value.length === opts.length ? [] : value,
-      }))
-      return
-    }
-    // select one
-    if (opts.includes(value)) {
-      setFilter((prev) => ({
-        ...prev,
-        [type]: opts.filter((v) => v !== value),
-      }))
-    } else {
-      setFilter((prev) => ({
-        ...prev,
-        [type]: [...opts, value],
-      }))
-    }
-  }
+  const { filterGauges, setSearchText, filter, onFilterChange } = useFilter(gauges)
 
   const gaugesTable = isDesktop ? (
     <AddGaugesTable
@@ -123,7 +49,7 @@ export const AddGaugeModal = ({ isOpen, onDismiss, selectRows, onGaugeAdd }) => 
       selectRows={selectRows}
       onRowSelect={onGaugeAdd}
       totalGaugesWeight={Number(totalGaugesWeight)}
-      data={filterRows}
+      data={filterGauges}
       maxHeight={70 * 5}
     />
   ) : (
@@ -132,21 +58,10 @@ export const AddGaugeModal = ({ isOpen, onDismiss, selectRows, onGaugeAdd }) => 
       selectRows={selectRows}
       onRowSelect={onGaugeAdd}
       totalGaugesWeight={Number(totalGaugesWeight)}
-      data={filterRows}
+      data={filterGauges}
       listDisplay="card"
     />
   )
-
-  useEffect(() => {
-    if (!isOpen) {
-      setSearchText('')
-      setFilter({
-        byChain: [],
-        byFeeTier: [],
-        byType: [],
-      })
-    }
-  }, [isOpen])
 
   return (
     <>
@@ -177,7 +92,7 @@ export const AddGaugeModal = ({ isOpen, onDismiss, selectRows, onGaugeAdd }) => 
                 </Button>
               </FlexGap>
               <Grid gridTemplateColumns={isDesktop ? '1fr 1fr' : '1fr'} gridGap={isDesktop ? '32px' : '1em'}>
-                <FilterFieldByType onFilterChange={setOption} />
+                <FilterFieldByType onFilterChange={onFilterChange} value={filter} />
                 <FilterFieldInput placeholder={t('Search gauges')} onChange={setSearchText} hideLabel={isMobile} />
               </Grid>
               {isMobile && selectRows?.length ? (
@@ -198,13 +113,6 @@ export const AddGaugeModal = ({ isOpen, onDismiss, selectRows, onGaugeAdd }) => 
           </Flex>
         </ModalWrapper>
       </ModalV2>
-      <FilterModal
-        isOpen={Boolean(option)}
-        onDismiss={() => setOption(null)}
-        type={option}
-        options={filter}
-        onChange={onFilterChange}
-      />
     </>
   )
 }
